+++ verticapy/learn/tools/__init__.py
@@ -14,4 +14,4 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.machine_learning.vertica.tools import *
+from verticapy.machine_learning.model_management.read import *
