+++ verticapy/machine_learning/_utils.py
@@ -20,6 +20,109 @@ from verticapy.sql._utils._format import quote_ident
 from verticapy._utils._sql import _executeSQL
 
 
+def _compute_tn_fn_fp_tp(
+    y_true: str,
+    y_score: str,
+    input_relation: Union[str, vDataFrame],
+    pos_label: Union[int, float, str] = 1,
+):
+    """
+A helper function that computes the confusion matrix for the specified 
+'pos_label' class and returns its values as a tuple of the following: 
+true negatives, false negatives, false positives, and true positives.
+
+Parameters
+----------
+y_true: str
+    Response column.
+y_score: str
+    Prediction.
+input_relation: str / vDataFrame
+    Relation to use for scoring. This relation can be a view, table, or a 
+    customized relation (if an alias is used at the end of the relation). 
+    For example: (SELECT ... FROM ...) x
+pos_label: int / float / str, optional
+    To compute the Confusion Matrix, one of the response column classes must 
+    be the positive one. The parameter 'pos_label' represents this class.
+
+Returns
+-------
+tuple
+    tn, fn, fp, tp
+    """
+    from verticapy.verticapy.machine_learning.metrics.classification import (
+        confusion_matrix,
+    )
+
+    matrix = confusion_matrix(y_true, y_score, input_relation, pos_label)
+    non_pos_label = 0 if (pos_label == 1) else f"Non-{pos_label}"
+    tn, fn, fp, tp = (
+        matrix.values[non_pos_label][0],
+        matrix.values[non_pos_label][1],
+        matrix.values[pos_label][0],
+        matrix.values[pos_label][1],
+    )
+    return tn, fn, fp, tp
+
+
+def _compute_metric_query(
+    metric: str,
+    y_true: str,
+    y_score: str,
+    input_relation: Union[str, vDataFrame],
+    title: str = "",
+    fetchfirstelem: bool = True,
+):
+    """
+A helper function that uses a specified metric to generate and score a query.
+
+Parameters
+----------
+metric: str
+    The metric to use in the query.
+y_true: str
+    Response column.
+y_score: str
+    Prediction.
+input_relation: str/vDataFrame
+    Relation to use for scoring. This relation can be a view, table, or a 
+    customized relation (if an alias is used at the end of the relation). 
+    For example: (SELECT ... FROM ...) x
+title: str, optional
+    Relation to use to do the scoring. The relation can be a view or a table
+    or even a customized relation. For example, you could write:
+    "(SELECT ... FROM ...) x" as long as an alias is given at the end of the
+Title of the query.
+fetchfirstelem: bool, optional
+    If set to True, this function returns one element. Otherwise, this 
+    function returns a tuple.
+
+Returns
+-------
+float or tuple of floats
+    score(s)
+    """
+    if isinstance(input_relation, str):
+        relation = input_relation
+    else:
+        relation = input_relation.__genSQL__()
+    if fetchfirstelem:
+        method = "fetchfirstelem"
+    else:
+        method = "fetchrow"
+    return _executeSQL(
+        query=f"""
+            SELECT 
+                /*+LABEL('learn.metrics._compute_metric_query')*/ 
+                {metric.format(y_true, y_score)} 
+            FROM {relation} 
+            WHERE {y_true} IS NOT NULL 
+              AND {y_score} IS NOT NULL;""",
+        title=title,
+        method=method,
+    )
+
+
 def compute_area(X: list, Y: list):
     auc = 0
     for i in range(len(Y) - 1):
