+++ verticapy/machine_learning/model_management/__init__.py
@@ -14,4 +14,7 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.sql.geo import *
+from verticapy.machine_learning.model_management.read import (
+    does_model_exist,
+    load_model,
+)
