+++ verticapy/machine_learning/model_management/read.py
@@ -27,11 +27,6 @@ from verticapy._version import vertica_version
 from verticapy.core.tablesample import tablesample
 from verticapy.sql._utils._format import quote_ident, schema_relation
 
-# Standard Python Modules
-import numpy as np
-from numpy.linalg import svd
-from typing import Union
-
 
 def does_model_exist(
     name: str, raise_error: bool = False, return_model_type: bool = False
@@ -683,58 +678,3 @@ model
                     }
                 )
     return model
-
-
-# This piece of code was taken from
-# https://en.wikipedia.org/wiki/Talk:Varimax_rotation
-
-
-def matrix_rotation(
-    Phi: Union[list, np.ndarray],
-    gamma: Union[int, float] = 1.0,
-    q: int = 20,
-    tol: float = 1e-6,
-):
-    """
-Performs a Oblimin (Varimax, Quartimax) rotation on the the model's 
-PCA matrix.
-
-Parameters
-----------
-Phi: list / numpy.array
-	input matrix.
-gamma: float, optional
-    Oblimin rotation factor, determines the type of rotation.
-    It must be between 0.0 and 1.0.
-        gamma = 0.0 results in a Quartimax rotation.
-        gamma = 1.0 results in a Varimax rotation.
-q: int, optional
-	Maximum number of iterations.
-tol: float, optional
-    The algorithm stops when the Frobenius norm of gradient is less than tol.
-
-Returns
--------
-model
-    The model.
-    """
-    Phi = np.array(Phi)
-    p, k = Phi.shape
-    R = np.eye(k)
-    d = 0
-    for i in range(q):
-        d_old = d
-        Lambda = np.dot(Phi, R)
-        u, s, vh = svd(
-            np.dot(
-                Phi.T,
-                np.asarray(Lambda) ** 3
-                - (gamma / p)
-                * np.dot(Lambda, np.diag(np.diag(np.dot(Lambda.T, Lambda)))),
-            )
-        )
-        R = np.dot(u, vh)
-        d = np.sum(s)
-        if d_old != 0 and d / d_old < 1 + tol:
-            break
-    return np.dot(Phi, R)
