+++ verticapy/sql/geo/__init__.py
@@ -14,4 +14,11 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.sql.geo import *
+from verticapy.sql.geo.geo import (
+    create_index,
+    coordinate_converter,
+    describe_index,
+    intersect,
+    rename_index,
+    split_polygon_n,
+)
