+++ verticapy/sql/geo/geo.py
